feat(gdb): Make gdb code less Linux-dependent
This is a squash of many intermediate, quite experimental steps,
which made rebasing essentially impossible.

- chore: stub/exclude macOS portions, fix errors
  - Use macOS-specific KickSignal
  - Fix some XhyveCpu implementations
  - Stub out/exclude unused gdb portions
  - Add placeholder AllBreakpoints for macOS
  - Add placeholder DebugExitInfo (may need to be replaced)

- refactor(gdb): Move run_gdb into crate::gdb
- refactor(vm): Move 'pub(crate) struct KickSignal' from 'os' to 'vm'
- refactor(treewide): Replace 'linux'/'macos' with 'os' for consistency
- fix(vcpu,bin): Fix cfg target attributes for MacOS
- chore(vm,bin/uhyve): Hide gdb-port cmdline option on MacOS
- docs(gdb): Add reference for IDET
- fix(gdb): Fix panic on try_into when passing exit code along
