+++ src/arch/x86_64/mod.rs
@@ -9,7 +9,7 @@ use x86_64::structures::paging::{
 	page_table::{FrameError, PageTableEntry},
 };
 
-use crate::{linux::x86_64::kvm_cpu::KVM_32BIT_GAP_START, mem::MmapMemory, paging::PagetableError};
+use crate::{mem::MmapMemory, os::x86_64::kvm_cpu::KVM_32BIT_GAP_START, paging::PagetableError};
 
 pub const PAGE_SIZE: usize = 0x1000;
 pub(crate) const GDT_OFFSET: u64 = 0x1000;
