+++ src/bin/uhyve.rs
@@ -187,9 +187,9 @@ struct UhyveArgs {
 	///
 	/// [default: 6677]
 	#[clap(short = 's', long, env = "HERMIT_GDB_PORT", num_args(0..=1), default_missing_value("6677"))]
+	#[cfg_attr(target_os = "macos", clap(hide(true)))]
 	#[serde(default)]
 	#[merge(strategy = merge::option::overwrite_none)]
-	#[cfg(target_os = "linux")]
 	gdb_port: Option<u16>,
 
 	/// TOML configuration file
@@ -463,7 +463,6 @@ impl From<Args> for Params {
 					file_isolation,
 					#[cfg(target_os = "linux")]
 					io_mode,
-					#[cfg(target_os = "linux")]
 					gdb_port,
 					config: _,
 					#[cfg(feature = "instrument")]
@@ -507,10 +506,7 @@ impl From<Args> for Params {
 			#[cfg(target_os = "linux")]
 			pit: pit.unwrap_or_default(),
 			file_mapping,
-			#[cfg(target_os = "linux")]
 			gdb_port,
-			#[cfg(target_os = "macos")]
-			gdb_port: None,
 			kernel_args,
 			tempdir,
 			hermit_image_mode: if let Some(hermit_image_mode) = hermit_image_mode {
@@ -728,7 +724,6 @@ mod tests {
 				file_isolation: None,
 				#[cfg(target_os = "linux")]
 				io_mode: None,
-				#[cfg(target_os = "linux")]
 				gdb_port: None,
 				config: Some(PathBuf::from("config.txt")),
 				#[cfg(feature = "instrument")]
@@ -798,7 +793,6 @@ mod tests {
 				file_isolation: Some(String::from("strict")),
 				#[cfg(target_os = "linux")]
 				io_mode: Some(String::from("direct")),
-				#[cfg(target_os = "linux")]
 				gdb_port: Some(1),
 				config: Some(PathBuf::from("config.txt")),
 				#[cfg(feature = "instrument")]
