+++ src/macos/aarch64/vcpu.rs
@@ -15,8 +15,8 @@ use crate::{
 		mair, tcr_size,
 	},
 	hypercall,
-	macos::aarch64::virtio_device::XHyveVirtioNetDevice,
 	mem::MmapMemory,
+	os::aarch64::virtio_device::XHyveVirtioNetDevice,
 	params::{NetworkMode, Params},
 	stats::CpuStats,
 	vcpu::{VcpuStopReason, VirtualCPU},
@@ -96,19 +96,12 @@ pub struct XhyveCpu {
 	kernel_info: Arc<KernelInfo>,
 	stats: Option<CpuStats>,
 }
-unsafe impl Send for XhyveCpu {}
 
-impl XhyveCpu {
-	pub fn get_root_pagetable(&self) -> GuestPhysAddr {
-		GuestPhysAddr::new(
-			self.vcpu
-				.as_ref()
-				.unwrap()
-				.read_system_register(SystemRegister::TTBR0_EL1)
-				.unwrap(),
-		)
-	}
-}
+/// `XhyveCpu` can be sent across threads as long as `thread_local_init` wasn't called yet.
+/// This is necessary during initialization.
+///
+/// TODO(fogti): Refactor to use type-state pattern instead.
+unsafe impl Send for XhyveCpu {}
 
 impl VirtualCPU for XhyveCpu {
 	fn thread_local_init(&mut self) -> HypervisorResult<()> {
@@ -337,6 +330,28 @@ impl VirtualCPU for XhyveCpu {
 	fn get_cpu_frequency(&self) -> Option<NonZero<u32>> {
 		warn!("CPU base frequency detection not implemented!");
 		None
+	}
+
+	fn get_root_pagetable(&self) -> GuestPhysAddr {
+		GuestPhysAddr::new(
+			self.vcpu
+				.as_ref()
+				.unwrap()
+				.read_system_register(SystemRegister::TTBR0_EL1)
+				.unwrap(),
+		)
+	}
+
+	fn get_vcpu_id(&self) -> usize {
+		self.id.try_into().unwrap()
+	}
+
+	fn apply_current_guest_debug(
+		&mut self,
+		_breakpoints: &crate::os::Breakpoints,
+		_resume_mode: crate::gdb::resume::ResumeMode,
+	) -> HypervisorResult<()> {
+		todo!()
 	}
 }
