+++ src/macos/gdb/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod breakpoints {
+	/// TODO: Implement breakpoinst for macOS.
+	#[derive(Clone, Debug, Default)]
+	pub struct AllBreakpoints;
+}
